third_party/blink/web_tests/virtual/pna-iframes-warning/external/wpt/fetch/private-network-access/iframe.tentative.https.window_include=from-public-expected.txt

This is a testharness.js-based test.
FAIL public to local: failed preflight. assert_equals: expected "timeout" but got "loaded"
FAIL public to local: missing CORS headers. assert_equals: expected "timeout" but got "loaded"
FAIL public to local: missing PNA header. assert_equals: expected "timeout" but got "loaded"
PASS public to local: success.
FAIL public to private: failed preflight. assert_equals: expected "timeout" but got "loaded"
FAIL public to private: missing CORS headers. assert_equals: expected "timeout" but got "loaded"
FAIL public to private: missing PNA header. assert_equals: expected "timeout" but got "loaded"
PASS public to private: success.
PASS public to public: no preflight required.
Harness: the test ran to completion.
